requirements.txt
python-aesel-client==0.3
